done nmrview
