+++ core/fslibs/parsing_routines/__init__.py
@@ -3,5 +3,5 @@ from core.fslibs.parsing_routines.ansig import parse_ansig_peaklist as ansig
 #from core.fslibs.parsing_routines.cara_simple import cara_simple
 from core.fslibs.parsing_routines.ccpnmrv2 import parse_ccpnmrv2_peaklist as ccpnmrv2
 from core.fslibs.parsing_routines.nmrdraw import parse_nmrdraw_peaklist as nmrdraw
-#from core.fslibs.parsing_routines.nmrview import nmrview
+from core.fslibs.parsing_routines.nmrview import parse_nmrview_peaklist as nmrview
 from core.fslibs.parsing_routines.sparky import parse_sparky_peaklist as sparky
